[REFACTOR/#43] ResponseMonthlyDiaryDto 파일명 및 구조 수정
